move shared aws helpers to utils package
