+++ lib/srv/alpnproxy/local_proxy.go
@@ -34,8 +34,8 @@ import (
 
 	"github.com/gravitational/teleport/lib/client"
 	"github.com/gravitational/teleport/lib/srv/alpnproxy/common"
-	appaws "github.com/gravitational/teleport/lib/srv/app/aws"
 	"github.com/gravitational/teleport/lib/utils"
+	"github.com/gravitational/teleport/lib/utils/aws"
 )
 
 // LocalProxy allows upgrading incoming connection to TLS where custom TLS values are set SNI ALPN and
@@ -323,7 +323,7 @@ func (l *LocalProxy) StartAWSAccessProxy(ctx context.Context) error {
 		Transport: tr,
 	}
 	err := http.Serve(l.cfg.Listener, http.HandlerFunc(func(rw http.ResponseWriter, req *http.Request) {
-		if err := appaws.VerifyAWSSignature(req, l.cfg.AWSCredentials); err != nil {
+		if err := aws.VerifyAWSSignature(req, l.cfg.AWSCredentials); err != nil {
 			log.WithError(err).Errorf("AWS signature verification failed.")
 			rw.WriteHeader(http.StatusForbidden)
 			return
