+++ lib/web/ui/app.go
@@ -22,7 +22,7 @@ import (
 
 	"github.com/gravitational/teleport/api/types"
 	"github.com/gravitational/teleport/lib/tlsca"
-	"github.com/gravitational/teleport/lib/utils"
+	"github.com/gravitational/teleport/lib/utils/aws"
 )
 
 // App describes an application
@@ -44,7 +44,7 @@ type App struct {
 	// AWSConsole if true, indicates that the app represents AWS management console.
 	AWSConsole bool `json:"awsConsole"`
 	// AWSRoles is a list of AWS IAM roles for the application representing AWS console.
-	AWSRoles []utils.AWSRole `json:"awsRoles,omitempty"`
+	AWSRoles []aws.AWSRole `json:"awsRoles,omitempty"`
 }
 
 // MakeAppsConfig contains parameters for converting apps to UI representation.
@@ -88,7 +88,7 @@ func MakeApps(c MakeAppsConfig) []App {
 		}
 
 		if teleApp.IsAWSConsole() {
-			app.AWSRoles = utils.FilterAWSRoles(c.Identity.AWSRoleARNs,
+			app.AWSRoles = aws.FilterAWSRoles(c.Identity.AWSRoleARNs,
 				teleApp.GetAWSAccountID())
 		}
